+++ MainFrame/Payroll/paytimeSheet/paytimeSheet.py
@@ -1,13 +1,12 @@
-import logging
 import sys
 import os
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
 from MainFrame.Resources.lib import *
-from MainFrame.TimeKeeping.payroll_functions.deductionFunctions import DeductionUI
+from MainFrame.Payroll.payroll_functions.deductionFunctions import DeductionUI
 from MainFrame.systemFunctions import globalFunction
-from MainFrame.TimeKeeping.payroll_functions.paytimeSheetFunctions import PaytimeSheetUI
+from MainFrame.Payroll.payroll_functions.paytimeSheetFunctions import PaytimeSheetUI
 from MainFrame.systemFunctions import timekeepingFunction
 
 warnings.filterwarnings("ignore", category=DeprecationWarning, message=".*sipPyTypeDict.*")
