+++ output/MainWindow/MainFrame/TimeKeeping/payTrans/payTransLoader.py
@@ -3,7 +3,7 @@ import os
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from MainFrame.Resources.lib import *
 from MainFrame.systemFunctions import globalFunction, timekeepingFunction
-from MainFrame.TimeKeeping.payTrans.payTransMailer import EmailerLoader
+from MainFrame.Payroll.payTrans.payTransMailer import EmailerLoader
 from MainFrame.Database_Connection.DBConnection import create_connection
 warnings.filterwarnings("ignore", category=DeprecationWarning, message=".*sipPyTypeDict.*")
