+++ output/MainWindow/MainFrame/TimeKeeping/paytimeSheet/paytimeSheet.py
@@ -1,12 +1,11 @@
-import logging
 import sys
 import os
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from MainFrame.Resources.lib import *
-from MainFrame.TimeKeeping.payTrans.payTransLoader import PayTrans
+from MainFrame.Payroll.payTrans.payTransLoader import PayTrans
 from MainFrame.systemFunctions import globalFunction, timekeepingFunction
-from MainFrame.TimeKeeping.paytimeSheet.storeDeductionLoader import StoreDeductionLoader
+from MainFrame.Payroll.paytimeSheet.storeDeductionLoader import StoreDeductionLoader
 from MainFrame.Database_Connection.user_session import UserSession
 
 warnings.filterwarnings("ignore", category=DeprecationWarning, message=".*sipPyTypeDict.*")
